requirements.txt
discord.py==2.4.0
gql==3.5.0
Pillow==10.4.0
python-dotenv==1.0.1
